Pundit

+++ app/controllers/cocktails_controller.rb
@@ -1,25 +1,26 @@
 class CocktailsController < ApplicationController
 skip_before_action :authenticate_user!, only: :index, raise: false
+before_action :set_cocktails, only: [ :show, :edit, :update, :destroy ]
   def index
     if params[:query].present?
-      @cocktails = Cocktail.where("name ILIKE ?", "%#{params[:query]}%")
+      @cocktails = policy_scope(Cocktail).where("name ILIKE ?", "%#{params[:query]}%")
 
     else
-      @cocktails = Cocktail.all
+      @cocktails = policy_scope(Cocktail)
     end
   end
 
 
   def show
-    @cocktail = Cocktail.find(params[:id])
     @dose = Dose.new
   end
 
   def create
-    authorize @cocktail
     @cocktail = Cocktail.new(cocktail_params)
+    @cocktail.user = current_user
+    authorize @cocktail
     if @cocktail.save
-      redirect_to cocktails_path
+      redirect_to cocktails_path, notice: "#{@cocktail.name} created. "
     else
       render 'cocktails/show'
     end
@@ -28,18 +29,19 @@ skip_before_action :authenticate_user!, only: :index, raise: false
 
   def new
     @cocktail = Cocktail.new
+    authorize @cocktail
   end
 
   def edit
-    @cocktail = Cocktail.find(params[:id])
     @dose = Dose.new
+
   end
 
   def update
-    @cocktail = Cocktail.find(params[:id])
     @dose = Dose.new
+    authorize @cocktail
     if @cocktail.update(cocktail_params)
-      redirect_to cocktails_path
+      redirect_to cocktails_path, notice: "You succesfully updated #{@cocktail.name} at #{Time.now}."
     else
       render :new
     end
@@ -47,8 +49,8 @@ skip_before_action :authenticate_user!, only: :index, raise: false
   end
 
   def destroy
-    @cocktail = Cocktail.find(params[:id])
     @dose = Dose.new
+    authorize @cocktail
     if @cocktail.destroy
       redirect_to cocktails_path, notice: "You removed #{@cocktail.name} which you created at #{@cocktail.created_at}."
     else
@@ -57,6 +59,11 @@ skip_before_action :authenticate_user!, only: :index, raise: false
   end
 
   private
+
+  def set_cocktails
+    @cocktail = Cocktail.find(params[:id])
+    authorize @cocktail
+  end
 
   def cocktail_params
     params.require(:cocktail).permit(:name, :photo)
